IO::File::Stream: fixed an annoying warning from gcc
git-svn-id: file:///home/michael/svn/yuni/trunk@2257 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/io/file/stream.hxx
@@ -251,7 +251,7 @@ namespace File
 	Stream::read(CString<CSizeT, ExpT>& buffer, uint64 size)
 	{
 		assert(pFd && "File not opened");
-		assert(size <= 2 * 1024 * 1024 * 1024);
+		assert(size <= uint64(2 * 1024) * 1024 * 1024);
 		if (!size)
 			return 0;
